shuffling around types
